+++ crates/rust-client/src/store/sqlite_store/account.rs
@@ -234,10 +234,16 @@ pub(super) fn insert_account_record(
     let account_seed = account_seed.map(|seed| seed.to_bytes());
 
     const QUERY: &str = "INSERT OR REPLACE INTO accounts (id, code_root, storage_root, vault_root, nonce, committed, account_seed, account_hash, locked) VALUES (?, ?, ?, ?, ?, ?, ?, ?, false)";
-    tx.execute(
-        QUERY,
-        params![id, code_root, storage_root, vault_root, nonce, committed, account_seed, hash],
-    )?;
+    tx.execute(QUERY, params![
+        id,
+        code_root,
+        storage_root,
+        vault_root,
+        nonce,
+        committed,
+        account_seed,
+        hash
+    ])?;
     Ok(())
 }